Add TestVideoPlayer project shell
