+++ Cpp.sln
@@ -174,6 +174,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VideoPlayerPrototype", "Vid
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DxPlayer", "VideoPlayer\DxPlayer\DxPlayer.vcxproj", "{1910C617-4F81-45C9-880C-379B079C8378}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "TestVideoPlayer", "VideoPlayer\TestVideoPlayer\TestVideoPlayer.vcxproj", "{490192C4-9071-49C1-97BB-7194C29E6AFF}"
+EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "VSIX", "VSIX", "{5E1FE3CC-1775-4DDD-835D-D3FEC7582CC7}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestCodeAnalyzer", "VSIX\TestCodeAnalyzer\TestCodeAnalyzer.csproj", "{419410D1-3EA3-4D11-940B-A81B342FE616}"
@@ -656,6 +658,24 @@ Global
 		{1910C617-4F81-45C9-880C-379B079C8378}.Release|x86.ActiveCfg = Release|Win32
 		{1910C617-4F81-45C9-880C-379B079C8378}.Release|x86.Build.0 = Release|Win32
 		{1910C617-4F81-45C9-880C-379B079C8378}.Release|x86.Deploy.0 = Release|Win32
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Debug|Any CPU.Build.0 = Debug|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Debug|Any CPU.Deploy.0 = Debug|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Debug|x64.ActiveCfg = Debug|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Debug|x64.Build.0 = Debug|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Debug|x64.Deploy.0 = Debug|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Debug|x86.ActiveCfg = Debug|Win32
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Debug|x86.Build.0 = Debug|Win32
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Debug|x86.Deploy.0 = Debug|Win32
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Release|Any CPU.ActiveCfg = Release|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Release|Any CPU.Build.0 = Release|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Release|Any CPU.Deploy.0 = Release|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Release|x64.ActiveCfg = Release|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Release|x64.Build.0 = Release|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Release|x64.Deploy.0 = Release|x64
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Release|x86.ActiveCfg = Release|Win32
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Release|x86.Build.0 = Release|Win32
+		{490192C4-9071-49C1-97BB-7194C29E6AFF}.Release|x86.Deploy.0 = Release|Win32
 		{419410D1-3EA3-4D11-940B-A81B342FE616}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{419410D1-3EA3-4D11-940B-A81B342FE616}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{419410D1-3EA3-4D11-940B-A81B342FE616}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -871,6 +891,7 @@ Global
 		{ADE9BE58-DBD7-4A0B-9273-4C59F7E77513} = {17F53077-EC87-44FA-AAC1-CC77982C25DE}
 		{49549E51-587F-4332-859C-5A1A64F69959} = {A0D9AFE0-98E0-4929-BDA6-785EF2113128}
 		{1910C617-4F81-45C9-880C-379B079C8378} = {A0D9AFE0-98E0-4929-BDA6-785EF2113128}
+		{490192C4-9071-49C1-97BB-7194C29E6AFF} = {A0D9AFE0-98E0-4929-BDA6-785EF2113128}
 		{419410D1-3EA3-4D11-940B-A81B342FE616} = {5E1FE3CC-1775-4DDD-835D-D3FEC7582CC7}
 		{74175162-7B5D-473F-9C88-B849B4768278} = {5E1FE3CC-1775-4DDD-835D-D3FEC7582CC7}
 		{89FEF167-68D4-46BB-BEBA-64948C4B6314} = {D8B16DFF-F06A-4548-82B2-67E2176BB7BF}
