--- VideoPlayer/DxPlayer_copy/App.xaml.h
@@ -1,28 +1,0 @@
-﻿//
-// App.xaml.h
-// Declaration of the App class.
-//
-
-#pragma once
-
-#include "App.g.h"
-#include "DirectXPage.xaml.h"
-
-namespace DxPlayer
-{
-		/// <summary>
-	/// Provides application-specific behavior to supplement the default Application class.
-	/// </summary>
-	ref class App sealed
-	{
-	public:
-		App();
-		virtual void OnLaunched(Windows::ApplicationModel::Activation::LaunchActivatedEventArgs^ e) override;
-
-	private:
-		void OnSuspending(Platform::Object^ sender, Windows::ApplicationModel::SuspendingEventArgs^ e);
-		void OnResuming(Platform::Object ^sender, Platform::Object ^args);
-		void OnNavigationFailed(Platform::Object ^sender, Windows::UI::Xaml::Navigation::NavigationFailedEventArgs ^e);
-		DirectXPage^ m_directXPage;
-	};
-}
