--- VideoPlayer/DxPlayer_copy/Common/Config.h
@@ -1,8 +1,0 @@
-﻿#pragma once
-#define DX_PLAYER_RENDER 1
-#define MEDIA_ENDIGNE_LOGIC 2
-#define MEDIA_ENDIGNE_COMPONENT 3
-
-//#define ENGINE_TYPE DX_PLAYER_RENDER
-#define ENGINE_TYPE MEDIA_ENDIGNE_LOGIC
-//#define ENGINE_TYPE MEDIA_ENDIGNE_COMPONENT
