--- VideoPlayer/DxPlayer_copy/DirectXPage.xaml.h
@@ -1,162 +1,0 @@
-﻿#pragma once
-#include "DirectXPage.g.h"
-
-#include "Common\Config.h"
-#include "DxPlayerMain.h"
-
-#if ENGINE_TYPE == MEDIA_ENDIGNE_COMPONENT
-namespace DxPlayer
-{
-	/// <summary>
-	/// A page that hosts a DirectX SwapChainPanel.
-	/// </summary>
-	public ref class DirectXPage sealed
-	{
-	public:
-		DirectXPage();
-		virtual ~DirectXPage();
-
-		void SaveInternalState(Windows::Foundation::Collections::IPropertySet^ state);
-		void LoadInternalState(Windows::Foundation::Collections::IPropertySet^ state);
-
-	private:
-		// XAML low-level rendering event handler.
-		void OnRendering(Platform::Object^ sender, Platform::Object^ args);
-
-		// Window event handlers.
-		void OnVisibilityChanged(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::VisibilityChangedEventArgs^ args);
-
-		// DisplayInformation event handlers.
-		void OnDpiChanged(Windows::Graphics::Display::DisplayInformation^ sender, Platform::Object^ args);
-		void OnOrientationChanged(Windows::Graphics::Display::DisplayInformation^ sender, Platform::Object^ args);
-		void OnDisplayContentsInvalidated(Windows::Graphics::Display::DisplayInformation^ sender, Platform::Object^ args);
-
-		// Other event handlers.
-		void AppBarButton_Click(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e);
-		void OnCompositionScaleChanged(Windows::UI::Xaml::Controls::SwapChainPanel^ sender, Object^ args);
-		void OnSwapChainPanelSizeChanged(Platform::Object^ sender, Windows::UI::Xaml::SizeChangedEventArgs^ e);
-
-		// Track our independent input on a background worker thread.
-		Windows::Foundation::IAsyncAction^ inputLoopWorker;
-		Windows::UI::Core::CoreIndependentInputSource^ coreInput;
-
-		// Independent input handling functions.
-		void OnPointerPressed(Platform::Object^ sender, Windows::UI::Core::PointerEventArgs^ e);
-		void OnPointerMoved(Platform::Object^ sender, Windows::UI::Core::PointerEventArgs^ e);
-		void OnPointerReleased(Platform::Object^ sender, Windows::UI::Core::PointerEventArgs^ e);
-
-		// Resources used to render the DirectX content in the XAML page background.
-		MediaEngineWinRT::Player^ playerWinRt;
-
-		bool windowVisible;
-	};
-}
-
-
-#elif ENGINE_TYPE == MEDIA_ENDIGNE_LOGIC
-#include <Helpers/Dx/SwapChainPanel.h>
-
-namespace DxPlayer
-{
-	/// <summary>
-	/// A page that hosts a DirectX SwapChainPanel.
-	/// </summary>
-	public ref class DirectXPage sealed
-	{
-	public:
-		DirectXPage();
-		virtual ~DirectXPage();
-
-		void SaveInternalState(Windows::Foundation::Collections::IPropertySet^ state);
-		void LoadInternalState(Windows::Foundation::Collections::IPropertySet^ state);
-
-	private:
-		Helpers::WinRt::Dx::SwapChainPanel^ DirectXPage::CreateSwapChainPanelWinRt(Windows::UI::Xaml::Controls::SwapChainPanel^ swapChainPanelXaml);
-
-		// XAML low-level rendering event handler.
-		void OnRendering(Platform::Object^ sender, Platform::Object^ args);
-
-		// Window event handlers.
-		void OnVisibilityChanged(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::VisibilityChangedEventArgs^ args);
-
-		// DisplayInformation event handlers.
-		void OnDpiChanged(Windows::Graphics::Display::DisplayInformation^ sender, Platform::Object^ args);
-		void OnOrientationChanged(Windows::Graphics::Display::DisplayInformation^ sender, Platform::Object^ args);
-		void OnDisplayContentsInvalidated(Windows::Graphics::Display::DisplayInformation^ sender, Platform::Object^ args);
-
-		// Other event handlers.
-		void AppBarButton_Click(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e);
-		void OnCompositionScaleChanged(Windows::UI::Xaml::Controls::SwapChainPanel^ sender, Object^ args);
-		void OnSwapChainPanelSizeChanged(Platform::Object^ sender, Windows::UI::Xaml::SizeChangedEventArgs^ e);
-
-		// Track our independent input on a background worker thread.
-		Windows::Foundation::IAsyncAction^ inputLoopWorker;
-		Windows::UI::Core::CoreIndependentInputSource^ coreInput;
-
-		// Independent input handling functions.
-		void OnPointerPressed(Platform::Object^ sender, Windows::UI::Core::PointerEventArgs^ e);
-		void OnPointerMoved(Platform::Object^ sender, Windows::UI::Core::PointerEventArgs^ e);
-		void OnPointerReleased(Platform::Object^ sender, Windows::UI::Core::PointerEventArgs^ e);
-
-
-	private:
-		// Resources used to render the DirectX content in the XAML page background.
-		Helpers::WinRt::Dx::SwapChainPanel^ swapChainPanelWinRt;
-
-		std::unique_ptr<DxPlayerMain> main; 
-		bool windowVisible;
-	};
-}
-
-#elif ENGINE_TYPE == DX_PLAYER_RENDER
-#include "Common\DeviceResources.h"
-
-namespace DxPlayer
-{
-	/// <summary>
-	/// A page that hosts a DirectX SwapChainPanel.
-	/// </summary>
-	public ref class DirectXPage sealed
-	{
-	public:
-		DirectXPage();
-		virtual ~DirectXPage();
-
-		void SaveInternalState(Windows::Foundation::Collections::IPropertySet^ state);
-		void LoadInternalState(Windows::Foundation::Collections::IPropertySet^ state);
-
-	private:
-		// XAML low-level rendering event handler.
-		void OnRendering(Platform::Object^ sender, Platform::Object^ args);
-
-		// Window event handlers.
-		void OnVisibilityChanged(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::VisibilityChangedEventArgs^ args);
-
-		// DisplayInformation event handlers.
-		void OnDpiChanged(Windows::Graphics::Display::DisplayInformation^ sender, Platform::Object^ args);
-		void OnOrientationChanged(Windows::Graphics::Display::DisplayInformation^ sender, Platform::Object^ args);
-		void OnDisplayContentsInvalidated(Windows::Graphics::Display::DisplayInformation^ sender, Platform::Object^ args);
-
-		// Other event handlers.
-		void AppBarButton_Click(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e);
-		void OnCompositionScaleChanged(Windows::UI::Xaml::Controls::SwapChainPanel^ sender, Object^ args);
-		void OnSwapChainPanelSizeChanged(Platform::Object^ sender, Windows::UI::Xaml::SizeChangedEventArgs^ e);
-
-		// Track our independent input on a background worker thread.
-		Windows::Foundation::IAsyncAction^ inputLoopWorker;
-		Windows::UI::Core::CoreIndependentInputSource^ coreInput;
-
-		// Independent input handling functions.
-		void OnPointerPressed(Platform::Object^ sender, Windows::UI::Core::PointerEventArgs^ e);
-		void OnPointerMoved(Platform::Object^ sender, Windows::UI::Core::PointerEventArgs^ e);
-		void OnPointerReleased(Platform::Object^ sender, Windows::UI::Core::PointerEventArgs^ e);
-
-		// Resources used to render the DirectX content in the XAML page background.
-		std::shared_ptr<DX::DeviceResources> deviceResources;
-	
-		std::unique_ptr<DxPlayerMain> main;
-		bool windowVisible;
-	};
-}
-
-#endif
