--- VideoPlayer/DxPlayer_copy/pch.h
@@ -1,16 +1,0 @@
-﻿#pragma once
-#include <wrl.h>
-#include <wrl/client.h>
-#include <dxgi1_4.h>
-#include <d3d11_3.h>
-#include <d2d1_3.h>
-#include <d2d1effects_2.h>
-#include <dwrite_3.h>
-#include <wincodec.h>
-#include <DirectXColors.h>
-#include <DirectXMath.h>
-#include <memory>
-#include <agile.h>
-#include <concrt.h>
-#include <collection.h>
-#include "App.xaml.h"
