+++ VideoPlayer/TestVideoPlayer/App.xaml
@@ -1,7 +1,6 @@
 ﻿<Application
-    x:Class="DxPlayer.App"
+    x:Class="TestVideoPlayer.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:DxPlayer">
-
+    xmlns:local="using:TestVideoPlayer">
 </Application>
