+++ VideoPlayer/TestVideoPlayer/MainPage.xaml.cpp
@@ -1,0 +1,11 @@
+﻿#include "pch.h"
+#include "MainPage.xaml.h"
+
+using namespace TestVideoPlayer;
+using namespace Platform;
+using namespace Windows::UI::Xaml;
+
+MainPage::MainPage()
+{
+	InitializeComponent();
+}
