+++ VideoPlayer/TestVideoPlayer/MainPage.xaml.h
@@ -1,0 +1,12 @@
+﻿#pragma once
+
+#include "MainPage.g.h"
+
+namespace TestVideoPlayer
+{
+	public ref class MainPage sealed
+	{
+	public:
+		MainPage();
+	};
+}
