+++ VideoPlayer/TestVideoPlayer/Package.appxmanifest
@@ -7,14 +7,14 @@
   IgnorableNamespaces="uap mp">
 
   <Identity
-    Name="DxPlayer"
+    Name="TestVideoPlayer"
     Publisher="CN=isrepeat"
     Version="1.0.0.0" />
 
   <mp:PhoneIdentity PhoneProductId="ab78d960-8d0c-49f6-b608-4aff1cfad103" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
 
   <Properties>
-    <DisplayName>DxPlayer</DisplayName>
+    <DisplayName>TestVideoPlayer</DisplayName>
     <PublisherDisplayName>CppTeam</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
@@ -30,12 +30,12 @@
   <Applications>
     <Application Id="App"
       Executable="$targetnametoken$.exe"
-      EntryPoint="DxPlayer.App">
+      EntryPoint="TestVideoPlayer.App">
       <uap:VisualElements
-        DisplayName="DxPlayer"
+        DisplayName="TestVideoPlayer"
         Square150x150Logo="Assets\Square150x150Logo.png"
         Square44x44Logo="Assets\Square44x44Logo.png"
-        Description="DxPlayer"
+        Description="TestVideoPlayer"
         BackgroundColor="transparent">
         <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png"/>
         <uap:SplashScreen Image="Assets\SplashScreen.png" />
