+++ VideoPlayer/TestVideoPlayer/pch.h
@@ -1,0 +1,6 @@
+﻿#pragma once
+
+#include <memory>
+#include <collection.h>
+#include <ppltasks.h>
+#include "App.xaml.h"
